requirements.txt
APScheduler==3.3.1
click==6.7
Flask==0.12.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
mysqlclient==1.3.10
PyAutoGUI==0.9.35
pycrypto==2.6.1
pytz==2017.2
requests==2.13.0
selenium==3.3.1
six==1.10.0
tzlocal==1.3
Werkzeug==0.12.1